fix: use common-lib

+++ auth-service/src/main/java/com/odms/auth/service/IAuthService.java
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.odms.auth.dto.request.LoginRequest;
 import com.odms.auth.dto.request.RegisterRequest;
 import com.odms.auth.dto.request.VerifyRequest;
-import com.odms.auth.dto.response.IDResponse;
 import com.odms.auth.dto.response.LoginResponse;
 import com.odms.auth.dto.response.VerifyResponse;
+import nmquan.commonlib.dto.response.IDResponse;
 
 public interface IAuthService {
     LoginResponse loginAccount(LoginRequest loginRequest);
     VerifyResponse verifyToken(VerifyRequest verifyRequest);
-    IDResponse<Integer> registerAccount(RegisterRequest request, String roleName) throws JsonProcessingException;
+    IDResponse<Long> registerAccount(RegisterRequest request, String roleName) throws JsonProcessingException;
     void verifyEmail(String token);
 }
