+++ auth-service/src/main/java/com/odms/auth/service/IDeliveryStaffService.java
@@ -1,12 +1,12 @@
 package com.odms.auth.service;
 
 import com.odms.auth.dto.response.DeliveryStaffResponse;
-import com.odms.auth.dto.response.IDResponse;
+import nmquan.commonlib.dto.response.IDResponse;
 
 import java.util.List;
 
 public interface IDeliveryStaffService {
     Boolean getMyStatusFindingOrder();
-    IDResponse<Integer> updateStatusFindingOrder(Integer userId);
+    IDResponse<Long> updateStatusFindingOrder(Long userId);
     List<DeliveryStaffResponse> findDeliveryStaff(Boolean status);
 }
