+++ auth-service/src/main/java/com/odms/auth/service/IUserService.java
@@ -1,14 +1,14 @@
 package com.odms.auth.service;
 
 import com.odms.auth.dto.request.FilterUserRequest;
-import com.odms.auth.dto.response.FilterResponse;
 import com.odms.auth.dto.response.UserResponse;
+import nmquan.commonlib.dto.response.FilterResponse;
 
 import java.util.List;
 import java.util.Map;
 
 public interface IUserService {
     UserResponse getCurrentUser();
-    Map<Integer, UserResponse> getUserByIds(List<Integer> ids);
+    Map<Long, UserResponse> getUserByIds(List<Long> ids);
     FilterResponse<UserResponse> filterUsers(FilterUserRequest request);
 }
